minor refactoring

+++ src/core/core/aio_provider.cpp
@@ -42,10 +42,7 @@ aio_provider::aio_provider(disk_engine *disk, aio_provider *inner_provider) : _e
 
 service_node *aio_provider::node() const { return _engine->node(); }
 
-void aio_provider::complete_io(aio_task *aio,
-                               error_code err,
-                               uint32_t bytes,
-                               int delay_milliseconds)
+void aio_provider::complete_io(aio_task *aio, error_code err, int64_t bytes, int delay_milliseconds)
 {
     _engine->complete_io(aio, err, bytes, delay_milliseconds);
 }
